Create sub-module for overwritten JIRA-http-client #43
