+++ jira-rest-java-client-fix/.gitignore
@@ -1,0 +1,8 @@
+/.settings/
+/src/main/lombok/lombok.config
+/src/test/lombok/lombok.config
+/src/lombok.config
+/target/
+/.checkstyle
+/.classpath
+/.project
