+++ jira-rest-java-client-fix/.travis.yml
@@ -1,0 +1,2 @@
+language: java
+install: mvn install -DskipTests=true -Dgpg.skip=true --batch-mode --show-version
