+++ jira-rest-java-client-fix/CHANGELOG.md
@@ -1,0 +1,4 @@
+# Changelog
+This package is part of [the Oversigt project](../README.md).
+
+All notable changes to this project will be documented in its [CHANGELOG.md](../CHANGELOG.md).
