+++ jira-rest-java-client-fix/README.md
@@ -1,0 +1,5 @@
+# JIRA REST Client Fix
+This package fixes the usage of Atlassians JIRA Rest Client Implementation by adding Fugue to the dependencies and improves the JIRA REST Client by adding a Guice injectable socket timeout named `jiraSocketTimeout`.
+
+## Getting started
+This package is part of [the Oversigt project](../README.md).
