+++ jira-rest-java-client-fix/pom.xml
@@ -1,0 +1,106 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>com.hlag.oversigt</groupId>
+		<artifactId>oversigt</artifactId>
+		<version>0.6-SNAPSHOT</version>
+	</parent>
+
+	<artifactId>jira-rest-java-client-fix</artifactId>
+
+	<name>JIRA REST Java Client Fix</name>
+	<url>http://github.com/Hapag-Lloyd/oversigt</url>
+	<organization>
+		<name>Hapag-Lloyd AG</name>
+		<url>https://www.hapag-lloyd.com/</url>
+	</organization>
+	<licenses>
+		<license>
+			<name>MIT License</name>
+			<url>https://opensource.org/licenses/MIT</url>
+		</license>
+	</licenses>
+
+	<developers>
+		<developer>
+			<name>Hapag-Lloyd AG</name>
+			<url>https://www.hapag-lloyd.com/</url>
+		</developer>
+	</developers>
+
+	<scm>
+		<connection>scm:git:git@github.com:Hapag-Lloyd/oversigt.git</connection>
+		<developerConnection>scm:git:git@github.com:Hapag-Lloyd/oversigt.git</developerConnection>
+		<url>https://github.com/Hapag-Lloyd/oversigt</url>
+	</scm>
+	<issueManagement>
+		<system>GitHub Issues</system>
+		<url>https://github.com/Hapag-Lloyd/oversigt/issues</url>
+	</issueManagement>
+	<ciManagement>
+		<system>Travis</system>
+		<url>https://travis-ci.org/Hapag-Lloyd/oversigt</url>
+	</ciManagement>
+
+	<dependencies>
+		<!-- JIRA REST Client Implementation -->
+		<dependency>
+			<groupId>com.atlassian.jira</groupId>
+			<artifactId>jira-rest-java-client-core</artifactId>
+			<version>5.1.0</version>
+		</dependency>
+
+		<!-- Missing inside the JIRA REST Client Implementation dependencies -->
+		<dependency>
+			<groupId>io.atlassian.fugue</groupId>
+			<artifactId>fugue</artifactId>
+			<version>4.7.2</version>
+		</dependency>
+
+		<!-- Dependency Injection -->
+		<dependency>
+			<groupId>com.google.inject</groupId>
+			<artifactId>guice</artifactId>
+			<version>4.2.1</version>
+		</dependency>
+	</dependencies>
+
+	<repositories>
+		<repository>
+			<id>atlassian-public</id>
+			<url>https://maven.atlassian.com/repository/public</url>
+		</repository>
+	</repositories>
+
+	<build>
+		<plugins>
+			<plugin>
+				<artifactId>maven-dependency-plugin</artifactId>
+				<executions>
+					<execution>
+						<!-- analyze-only -->
+						<id>default</id>
+						<configuration>
+							<ignoredUsedUndeclaredDependencies>
+								<!-- Dependencies of the JIRA REST Client Implementation -->
+								<ignoredUsedUndeclaredDependency>com.atlassian.event:atlassian-event:jar:3.0.0</ignoredUsedUndeclaredDependency>
+								<ignoredUsedUndeclaredDependency>com.atlassian.httpclient:atlassian-httpclient-api:jar:2.0.0-m1</ignoredUsedUndeclaredDependency>
+								<ignoredUsedUndeclaredDependency>com.atlassian.httpclient:atlassian-httpclient-library:jar:2.0.0-m1</ignoredUsedUndeclaredDependency>
+								<ignoredUsedUndeclaredDependency>com.atlassian.jira:jira-rest-java-client-api:jar:5.1.0</ignoredUsedUndeclaredDependency>
+								<ignoredUsedUndeclaredDependency>com.atlassian.sal:sal-api:jar:3.0.7</ignoredUsedUndeclaredDependency>
+								<ignoredUsedUndeclaredDependency>com.google.code.findbugs:jsr305:jar:3.0.2</ignoredUsedUndeclaredDependency>
+								<ignoredUsedUndeclaredDependency>org.slf4j:slf4j-api:jar:1.7.10</ignoredUsedUndeclaredDependency>
+							</ignoredUsedUndeclaredDependencies>
+							<usedDependencies>
+								<!-- Missing inside the JIRA REST Client Implementation dependencies -->
+								<usedDependency>io.atlassian.fugue:fugue</usedDependency>
+							</usedDependencies>
+						</configuration>
+					</execution>
+				</executions>
+			</plugin>
+		</plugins>
+	</build>
+</project>
