+++ jira-rest-java-client-fix/src/main/lombok/com/atlassian/jira/rest/client/internal/async/package-info.java
@@ -1,2 +1,5 @@
+/**
+ * JIRA REST Client internals for asynchronous things
+ */
 @de.larssh.utils.annotations.NonNullByDefault
 package com.atlassian.jira.rest.client.internal.async;
